# specs/001-intellistack-platform/contracts/learning.openapi.yaml
openapi: 3.1.0
info:
  title: IntelliStack Learning API
  version: 1.0.0
  description: Learning progress, prerequisites, badges, certificates (FR-001 to FR-015)
servers:
  - url: /api/v1/learning
    description: Learning API base path
...
paths:
...
  /stages/{stageId}:
    get:
      tags: [Stages]
      summary: Get stage details
      operationId: getStage
      parameters:
        - name: stageId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Stage details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StageDetailResponse'
        '404':
          description: Stage not found
  /progress:
    get:
      tags: [Progress]
      summary: Get current user's progress (FR-003)
      operationId: getProgress
      security:
        - sessionAuth: []
      responses:
        '200':
          description: User progress
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProgressResponse'
  /progress/content/{contentId}/complete:
    post:
      tags: [Progress]
      summary: Mark content as completed (FR-003, FR-008)
      operationId: completeContent
      security:
        - sessionAuth: []
      parameters:
        - name: contentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                timeSpent:
                  type: integer
                  description: Seconds spent on content
      responses:
        '200':
          description: Content marked complete
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContentCompletionResponse'
        '403':
          description: Prerequisites not met (FR-001)
  /progress/stages/{stageId}/unlock:
    get:
      tags: [Progress]
      summary: Check if stage is unlocked (FR-001)
      operationId: checkStageUnlock
      security:
        - sessionAuth: []
      parameters:
        - name: stageId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Unlock status
          content:
            application/json:
              schema:
                type: object
                properties:
                  unlocked:
                    type: boolean
                  reason:
                    type: string
                  missingPrerequisites:
                    type: array
                    items:
                      type: string
  /progress/path:
    get:
      tags: [Progress]
      summary: Get personalized learning path (FR-012)
      operationId: getLearningPath
      security:
        - sessionAuth: []
      responses:
        '200':
          description: Learning path visualization
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LearningPathResponse'
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: session
  schemas:
    StageResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        number:
          type: integer
        name:
          type: string
        description:
          type: string
        estimatedHours:
          type: integer
        badgeId:
          type: string
          format: uuid
    StageDetailResponse:
      allOf:
        - $ref: '#/components/schemas/StageResponse'
        - type: object
          properties:
            learningObjectives:
              type: array
              items:
                type: string
            prerequisites:
              type: array
              items:
                $ref: '#/components/schemas/StageResponse'
            content:
              type: array
              items:
                $ref: '#/components/schemas/ContentSummary'
    ContentSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        type:
          type: string
          enum: [lesson, exercise, simulation, resource]
        orderIndex:
          type: integer
    ProgressResponse:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        currentStage:
          $ref: '#/components/schemas/StageResponse'
        completedStages:
          type: array
          items:
            type: integer
        contentCompletions:
          type: array
          items:
            $ref: '#/components/schemas/ContentCompletionSummary'
        totalTimeOnTask:
          type: integer
          description: Total seconds
        challengePathwayUsed:
          type: boolean
        stagesSkipped:
          type: array
          items:
            type: integer
        startedAt:
          type: string
          format: date-time
    ContentCompletionSummary:
      type: object
      properties:
        contentId:
          type: string
          format: uuid
        completedAt:
          type: string
          format: date-time
        timeSpent:
          type: integer
    ContentCompletionResponse:
      type: object
      properties:
        contentId:
          type: string
          format: uuid
        completedAt:
          type: string
          format: date-time
        stageProgress:
          type: number
          format: float
          description: Percentage complete in current stage
        badgeEarned:
          $ref: '#/components/schemas/BadgeResponse'
        nextContent:
          $ref: '#/components/schemas/ContentSummary'
    LearningPathResponse:
      type: object
      properties:
        stages:
          type: array
          items:
            type: object
            properties:
              stage:
                $ref: '#/components/schemas/StageResponse'
              status:
                type: string
                enum: [completed, current, locked]
              progress:
                type: number
                format: float
...
    BadgeResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        imageUrl:
          type: string
          format: uri
        stageId:
          type: string
          format: uuid
...